GETSAFE-000 : Fix route issue for developer journey, move buy-flow inside pages and fix  alignments
