+++ src/App.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import logo from './logo.svg';
 import './App.css';
-import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-import Buyflow from './buyflow/Buyflow';
-import { ProductIds } from './config';
+import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom';
+import { BuyFlow } from './pages/buy-flow';
+import { ProductIds, Products } from './config';
 import { Home } from './pages/home';
 
 const App = () => {
@@ -11,16 +11,18 @@ const App = () => {
         <Router>
             <div className='App'>
                 <header className='App-header'>
-                    <img src={logo} className='App-logo' alt='logo' />
+                    <Link to='/'>
+                        <img src={logo} className='App-logo' alt='logo' />
+                    </Link>
                 </header>
                 <Switch>
-                    <Route path='/buy/insurance_dev'>
-                        <Buyflow productId={ProductIds.devIns} />
+                    <Route path={Products.dev_ins.url} exact>
+                        <BuyFlow productId={ProductIds.devIns} />
                     </Route>
-                    <Route path='/buy/insurance_design'>
-                        <Buyflow productId={ProductIds.designIns} />
+                    <Route path={Products.design_ins.url} exact>
+                        <BuyFlow productId={ProductIds.designIns} />
                     </Route>
-                    <Route path='/'>
+                    <Route path='/' exact>
                         <Home />
                     </Route>
                 </Switch>
