+++ src/config.ts
@@ -1,16 +1,16 @@
 export enum ProductIds {
-  devIns = 'dev_ins',
-  designIns = 'design_ins',
+    devIns = 'dev_ins',
+    designIns = 'design_ins'
 }
 
 export const Products = {
-  [ProductIds.devIns]: {
-    name: 'Developer Insurance',
-    url: '/buy/insurance_developer',
-  },
+    [ProductIds.devIns]: {
+        name: 'Developer Insurance',
+        url: '/buy/insurance_dev'
+    },
 
-  [ProductIds.designIns]: {
-    name: 'Designer Insurance',
-    url: '/buy/insurance_design',
-  },
-}
+    [ProductIds.designIns]: {
+        name: 'Designer Insurance',
+        url: '/buy/insurance_design'
+    }
+};
