+++ src/pages/buy-flow/AgeStep.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
-import { Input } from '../components/input';
-import { Button } from '../components/button';
+import { Input } from '../../components/input';
+import { Button } from '../../components/button';
 
 interface AgeStepProps {
     cb: (field: string, value: number) => void;
