+++ src/pages/buy-flow/index.ts
@@ -1,0 +1,1 @@
+export * from './buy-flow';
