+++ src/pages/home/home.css
@@ -6,7 +6,3 @@
 .insurance-list li {
     padding: 5px 0;
 }
-
-.home-page-text {
-    text-align: center;
-}
